requirements.txt
openai-whisper
torch
pyannote.audio
pydub
python-dotenv
